add graficos

+++ db/migrate/20201002211949_create_clientes.rb
@@ -2,7 +2,7 @@ class CreateClientes < ActiveRecord::Migration[5.1]
   def change
     create_table :clientes do |t|
       t.string :nome
-      t.integer :cpf
+      t.bigint :cpf
       t.string :telefone
       t.text :endereco
